+++ changelog.md
@@ -10,6 +10,8 @@
 - Merging players didn't merge their stage membership
 - Get display name from WW on login
 - width for extra rounds form
+- show assets field error on series create
+- If series recording access is GUILD any guild members can be added to game
 
 ## [1.12.10] - 2026-6-30 Small Bugs Need Squashing Too